tests_e2e/requirements.txt
pytest>=7.4.0
requests>=2.31.0
Pillow>=10.0.0
numpy>=1.24.0
